Move form CTA card to reusable component
