+++ next/src/components/page-contents/FormLandingPageContent/FormLandingPageContent.tsx
@@ -1,21 +1,17 @@
 import { Typography } from '@bratislava/component-library'
-import { useMutation } from '@tanstack/react-query'
-import { useRouter } from 'next/router'
 import { useTranslation } from 'next-i18next/pages'
 
-import { formsClient } from '@/src/clients/forms'
 import { FormWithLandingPageFragment } from '@/src/clients/graphql-strapi/api'
 import TableOfContents from '@/src/components/common/TableOfContents/TableOfContents'
 import Markdown from '@/src/components/formatting/Markdown'
 import { ClientLandingPageFormDefinition } from '@/src/components/forms/clientFormDefinitions'
 import SectionContainer from '@/src/components/layouts/SectionContainer'
 import Sections from '@/src/components/layouts/Sections'
+import FormLandingPageCtaCard from '@/src/components/page-contents/FormLandingPageContent/FormCta/FormLandingPageCtaCard'
 import FormLandingPageCard from '@/src/components/segments/FormLandingPageCard/FormLandingPageCard'
 import MLink from '@/src/components/simple-components/MLink'
-import useToast from '@/src/components/simple-components/Toast/useToast'
 import { isDefined } from '@/src/frontend/utils/general'
 import cn from '@/src/utils/cn'
-import { ROUTES } from '@/src/utils/routes'
 
 /**
  * Figma: https://www.figma.com/design/0VrrvwWs7n3T8YFzoHe92X/BK--Dizajn--DEV-?node-id=14475-7297
@@ -37,38 +33,7 @@ export type FormLandingPageProps = {
 }
 
 const FormLandingPage = ({ formDefinition, strapiForm }: FormLandingPageProps) => {
-  const router = useRouter()
   const { t } = useTranslation('forms')
-  const { showToast, closeToasts } = useToast()
-
-  const { mutate, isPending } = useMutation({
-    mutationFn: () =>
-      formsClient.formsV2ControllerCreateForm(
-        {
-          formDefinitionSlug: formDefinition.slug,
-        },
-        { authStrategy: 'authOrGuestWithToken' },
-      ),
-    networkMode: 'always',
-    onMutate: () => {
-      showToast({
-        message: t('form_landing_page.redirect_info'),
-        variant: 'info',
-        // Keep this toast visible for the whole redirect flow; it is closed explicitly after navigation succeeds.
-        duration: Number.MAX_SAFE_INTEGER,
-      })
-    },
-    onSuccess: async (response) => {
-      await router.push(
-        ROUTES.MUNICIPAL_SERVICES_FORM_WITH_ID(formDefinition.slug, response.data.formId),
-      )
-      // Close the redirect toast only after a successful route change so it stays visible until the user is redirected.
-      closeToasts()
-    },
-    onError: () => {
-      showToast({ message: t('form_landing_page.redirect_error'), variant: 'error' })
-    },
-  })
 
   const filteredSections = strapiForm.landingPage.sections?.filter(isDefined) ?? []
 
@@ -122,18 +87,15 @@ const FormLandingPage = ({ formDefinition, strapiForm }: FormLandingPageProps) =
                 <FormLandingPageCard key={linkCta.id} {...linkCta} />
               ))}
               {isDefined(strapiForm.landingPage.formCta) ? (
-                <FormLandingPageCard
-                  {...strapiForm.landingPage.formCta}
-                  isLoading={isPending}
-                  onPress={() => {
-                    mutate()
-                  }}
+                <FormLandingPageCtaCard
+                  formCta={strapiForm.landingPage.formCta}
+                  formDefinition={formDefinition}
                 />
               ) : null}
             </div>
           </SectionContainer>
         </div>
-        
+
         <aside className="w-full lg:top-40 lg:w-80 lg:shrink-0">
           <TableOfContents />
         </aside>
